Remove type checks from conflict check
These type checks belong in the `similar` and `equal` functions.
This makes this code much simpler, and the tests still pass. 

One test had to be corrected, in order to mark the `1` and `2` as similar, 
but this is an improvement - previously they were only 'similar' because they 
weren't objects or strings.